this commit introduces 1.0 of sampo.
There are considerable changes from what the MVP had (in order of impact/size of
the change):

- more of a plug-and-play experience for the end user.  sampo now requires a
  folder somewhere that contains: `sampo.sh`, `sampo.conf` and `scripts/`.  This
drastically simplifies deployments in both containers and local instances and
provides a nice workflow:
  - `mkdir -p sampo/scripts`
  - `cp yourscripts/*.sh sampo/scripts`
  - `cp sampo.sh sampo/`
  - `cp sampo.conf sampo/`
- this allows nice portability and a faster deployment.  The user dumps their
  scripts in `scripts/` then updates `sampo.conf` to set their endpoints and the
scripts that they call.  for containers, this can be done by adding them to the
dockerfile, using a volume mount, or config maps in k8s.  This should provide a
lot of flexibility for someone with a lot of scripts to dump them in one spot,
configure them, and test them
- more versatile container deployement on both docker and kubernetes with better
  logging and parity with local deployment
- `build.sh` now deploys local, docker, or k8s variants (`-l`, `-d`, `-k`)
- `build.sh` now cleans up previous deployments `-c`, and runs tests `./build
  -i`
- added debug logging
- more comments in the code
- enhanced some functions
- better portability of some commands like stat

since making this three years ago, I have seen the stars on the repo slowly
increase.  It is surpising to me, but nonetheless, it shows some interest, so I
decided to push the mvp to a better state.

If you are using sampo for something, I would like to hear from you on how you
are using it and what you would like to see from it.

thank you for using sampo.

Signed-off-by: Jacob Salmela <me@jacobsalmela.com>